Move FioAddress utils out of the modules folder
