+++ src/actions/SendConfirmationActions.tsx
@@ -9,13 +9,13 @@ import { InsufficientFeesModal } from '../components/modals/InsufficientFeesModa
 import { Airship, showError } from '../components/services/AirshipInstance'
 import { FEE_ALERT_THRESHOLD, FIO_STR } from '../constants/WalletAndCurrencyConstants'
 import { lstrings } from '../locales/strings'
-import { addToFioAddressCache, FIO_FEE_EXCEEDS_SUPPLIED_MAXIMUM, recordSend } from '../modules/FioAddress/util'
 import { initialState, initialTransaction } from '../reducers/scenes/SendConfirmationReducer'
 import { getExchangeDenomination } from '../selectors/DenominationSelectors'
 import { convertCurrency, convertCurrencyFromExchangeRates, getExchangeRate } from '../selectors/WalletSelectors'
 import { RootState, ThunkAction } from '../types/reduxTypes'
 import { NavigationBase } from '../types/routerTypes'
 import { GuiMakeSpendInfo, SpendAuthType } from '../types/types'
+import { addToFioAddressCache, FIO_FEE_EXCEEDS_SUPPLIED_MAXIMUM, recordSend } from '../util/FioAddressUtils'
 import { logActivity } from '../util/logger'
 import { convertNativeToExchange, DECIMAL_PRECISION, getDenomFromIsoCode } from '../util/utils'
 import { playSendSound } from './SoundActions'
