+++ src/components/FioAddress/ConnectWallets.tsx
@@ -3,17 +3,17 @@ import { EdgeAccount, EdgeCurrencyWallet } from 'edge-core-js'
 import * as React from 'react'
 import { ScrollView, Switch, View } from 'react-native'
 
-import { CryptoIcon } from '../../../components/icons/CryptoIcon'
-import { showError } from '../../../components/services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../../../components/services/ThemeContext'
-import { EdgeText } from '../../../components/themed/EdgeText'
-import { MainButton } from '../../../components/themed/MainButton'
-import { lstrings } from '../../../locales/strings'
-import { connect } from '../../../types/reactRedux'
-import { NavigationBase } from '../../../types/routerTypes'
-import { FioConnectionWalletItem } from '../../../types/types'
-import { getTokenId } from '../../../util/CurrencyInfoHelpers'
-import { convertFIOToEdgeCodes, makeConnectWallets } from '../util'
+import { CryptoIcon } from '../../components/icons/CryptoIcon'
+import { showError } from '../../components/services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../components/services/ThemeContext'
+import { EdgeText } from '../../components/themed/EdgeText'
+import { MainButton } from '../../components/themed/MainButton'
+import { lstrings } from '../../locales/strings'
+import { connect } from '../../types/reactRedux'
+import { NavigationBase } from '../../types/routerTypes'
+import { FioConnectionWalletItem } from '../../types/types'
+import { getTokenId } from '../../util/CurrencyInfoHelpers'
+import { convertFIOToEdgeCodes, makeConnectWallets } from '../../util/FioAddressUtils'
 
 interface LocalState {
   connectWalletsMap: { [walletId: string]: FioConnectionWalletItem }
