+++ src/components/FioAddress/DomainListModal.tsx
@@ -3,18 +3,18 @@ import * as React from 'react'
 import { TouchableOpacity, View } from 'react-native'
 import { AirshipBridge } from 'react-native-airship'
 
-import { Fontello } from '../../../assets/vector'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../../../components/services/ThemeContext'
-import { ClickableText } from '../../../components/themed/ClickableText'
-import { EdgeText } from '../../../components/themed/EdgeText'
-import { ModalFooter, ModalFooterFade, ModalTitle } from '../../../components/themed/ModalParts'
-import { OutlinedTextInput } from '../../../components/themed/OutlinedTextInput'
-import { ThemedModal } from '../../../components/themed/ThemedModal'
-import { FIO_ADDRESS_DELIMITER, FIO_DOMAIN_DEFAULT } from '../../../constants/WalletAndCurrencyConstants'
-import { lstrings } from '../../../locales/strings'
-import { connect } from '../../../types/reactRedux'
-import { NavigationBase } from '../../../types/routerTypes'
-import { FioDomain, FlatListItem } from '../../../types/types'
+import { Fontello } from '../../assets/vector'
+import { FIO_ADDRESS_DELIMITER, FIO_DOMAIN_DEFAULT } from '../../constants/WalletAndCurrencyConstants'
+import { lstrings } from '../../locales/strings'
+import { connect } from '../../types/reactRedux'
+import { NavigationBase } from '../../types/routerTypes'
+import { FioDomain, FlatListItem } from '../../types/types'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
+import { ClickableText } from '../themed/ClickableText'
+import { EdgeText } from '../themed/EdgeText'
+import { ModalFooter, ModalFooterFade, ModalTitle } from '../themed/ModalParts'
+import { OutlinedTextInput } from '../themed/OutlinedTextInput'
+import { ThemedModal } from '../themed/ThemedModal'
 
 interface Item {
   label: string
