+++ src/components/FioAddress/FioActionSubmit.tsx
@@ -3,20 +3,20 @@ import { EdgeCurrencyWallet } from 'edge-core-js'
 import * as React from 'react'
 import { ActivityIndicator, View } from 'react-native'
 
-import { WalletListModal, WalletListResult } from '../../../components/modals/WalletListModal'
-import { Airship, showError, showToast } from '../../../components/services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../../../components/services/ThemeContext'
-import { EdgeText } from '../../../components/themed/EdgeText'
-import { MainButton } from '../../../components/themed/MainButton'
-import { Slider } from '../../../components/themed/Slider'
-import { Tile } from '../../../components/tiles/Tile'
-import { FIO_STR } from '../../../constants/WalletAndCurrencyConstants'
-import { lstrings } from '../../../locales/strings'
-import { getDisplayDenomination } from '../../../selectors/DenominationSelectors'
-import { connect } from '../../../types/reactRedux'
-import { NavigationBase } from '../../../types/routerTypes'
-import { getAvailableBalance, getWalletName } from '../../../util/CurrencyWalletHelpers'
-import { DECIMAL_PRECISION, truncateDecimals } from '../../../util/utils'
+import { FIO_STR } from '../../constants/WalletAndCurrencyConstants'
+import { lstrings } from '../../locales/strings'
+import { getDisplayDenomination } from '../../selectors/DenominationSelectors'
+import { connect } from '../../types/reactRedux'
+import { NavigationBase } from '../../types/routerTypes'
+import { getAvailableBalance, getWalletName } from '../../util/CurrencyWalletHelpers'
+import { DECIMAL_PRECISION, truncateDecimals } from '../../util/utils'
+import { WalletListModal, WalletListResult } from '../modals/WalletListModal'
+import { Airship, showError, showToast } from '../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
+import { EdgeText } from '../themed/EdgeText'
+import { MainButton } from '../themed/MainButton'
+import { Slider } from '../themed/Slider'
+import { Tile } from '../tiles/Tile'
 
 type ActionResult =
   | {
