+++ src/components/scenes/Fio/FioAddressDetailsScene.tsx
@@ -4,11 +4,11 @@ import { Alert } from 'react-native'
 import IonIcon from 'react-native-vector-icons/Ionicons'
 
 import { lstrings } from '../../../locales/strings'
-import { ConnectWalletsConnector as ConnectWallets } from '../../../modules/FioAddress/components/ConnectWallets'
-import { BUNDLED_TXS_AMOUNT_ALERT, findWalletByFioAddress } from '../../../modules/FioAddress/util'
 import { connect } from '../../../types/reactRedux'
 import { EdgeSceneProps } from '../../../types/routerTypes'
+import { BUNDLED_TXS_AMOUNT_ALERT, findWalletByFioAddress } from '../../../util/FioAddressUtils'
 import { SceneWrapper } from '../../common/SceneWrapper'
+import { ConnectWalletsConnector as ConnectWallets } from '../../FioAddress/ConnectWallets'
 import { showError } from '../../services/AirshipInstance'
 import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
 import { SettingsHeaderRow } from '../../settings/SettingsHeaderRow'
