+++ src/components/scenes/Fio/FioAddressListScene.tsx
@@ -7,11 +7,11 @@ import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
 import fioAddressLogo from '../../../assets/images/fio/fio_logo.png'
 import { Fontello } from '../../../assets/vector'
 import { lstrings } from '../../../locales/strings'
-import { FioNameRow } from '../../../modules/FioAddress/components/FioName'
 import { connect } from '../../../types/reactRedux'
 import { EdgeSceneProps } from '../../../types/routerTypes'
 import { FioAddress, FioDomain } from '../../../types/types'
 import { SceneWrapper } from '../../common/SceneWrapper'
+import { FioNameRow } from '../../FioAddress/FioName'
 import { FullScreenLoader } from '../../progress-indicators/FullScreenLoader'
 import { showError } from '../../services/AirshipInstance'
 import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
