+++ src/components/scenes/Fio/FioAddressRegisterScene.tsx
@@ -7,14 +7,14 @@ import { createFioWallet } from '../../../actions/FioAddressActions'
 import { Fontello } from '../../../assets/vector'
 import { FIO_ADDRESS_DELIMITER, FIO_DOMAIN_DEFAULT } from '../../../constants/WalletAndCurrencyConstants'
 import { lstrings } from '../../../locales/strings'
-import { DomainListModal } from '../../../modules/FioAddress/components/DomainListModal'
-import { checkIsDomainPublic } from '../../../modules/FioAddress/util'
 import { connect } from '../../../types/reactRedux'
 import { EdgeSceneProps } from '../../../types/routerTypes'
 import { FioDomain, FioPublicDomain } from '../../../types/types'
 import { getWalletName } from '../../../util/CurrencyWalletHelpers'
+import { checkIsDomainPublic } from '../../../util/FioAddressUtils'
 import { openLink } from '../../../util/utils'
 import { SceneWrapper } from '../../common/SceneWrapper'
+import { DomainListModal } from '../../FioAddress/DomainListModal'
 import { TextInputModal } from '../../modals/TextInputModal'
 import { WalletListModal, WalletListResult } from '../../modals/WalletListModal'
 import { Airship, showError, showToast } from '../../services/AirshipInstance'
