+++ src/components/scenes/Fio/FioAddressSettingsScene.tsx
@@ -3,12 +3,12 @@ import * as React from 'react'
 
 import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
 import { lstrings } from '../../../locales/strings'
-import { FioActionSubmit } from '../../../modules/FioAddress/components/FioActionSubmit'
-import { addBundledTxs, getAddBundledTxsFee, getTransferFee } from '../../../modules/FioAddress/util'
 import { connect } from '../../../types/reactRedux'
 import { EdgeSceneProps } from '../../../types/routerTypes'
 import { GuiMakeSpendInfo } from '../../../types/types'
+import { addBundledTxs, getAddBundledTxsFee, getTransferFee } from '../../../util/FioAddressUtils'
 import { SceneWrapper } from '../../common/SceneWrapper'
+import { FioActionSubmit } from '../../FioAddress/FioActionSubmit'
 import { ButtonsModal } from '../../modals/ButtonsModal'
 import { Airship, showError, showToast } from '../../services/AirshipInstance'
 import { ThemeProps, withTheme } from '../../services/ThemeContext'
