+++ src/components/scenes/Fio/FioConnectWalletConfirmScene.tsx
@@ -2,11 +2,11 @@ import * as React from 'react'
 import { ScrollView } from 'react-native'
 
 import { lstrings } from '../../../locales/strings'
-import { FIO_NO_BUNDLED_ERR_CODE, updatePubAddressesForFioAddress } from '../../../modules/FioAddress/util'
 import { CcWalletMap } from '../../../reducers/FioReducer'
 import { connect } from '../../../types/reactRedux'
 import { EdgeSceneProps } from '../../../types/routerTypes'
 import { FioConnectionWalletItem } from '../../../types/types'
+import { FIO_NO_BUNDLED_ERR_CODE, updatePubAddressesForFioAddress } from '../../../util/FioAddressUtils'
 import { SceneWrapper } from '../../common/SceneWrapper'
 import { ButtonsModal } from '../../modals/ButtonsModal'
 import { Airship, showError, showToast } from '../../services/AirshipInstance'
