+++ src/components/scenes/Fio/FioDomainRegisterSelectWalletScene.tsx
@@ -7,7 +7,6 @@ import { sprintf } from 'sprintf-js'
 
 import { FIO_STR } from '../../../constants/WalletAndCurrencyConstants'
 import { lstrings } from '../../../locales/strings'
-import { getDomainRegInfo } from '../../../modules/FioAddress/util'
 import { getDisplayDenomination, getExchangeDenomination } from '../../../selectors/DenominationSelectors'
 import { config } from '../../../theme/appConfig'
 import { connect } from '../../../types/reactRedux'
@@ -15,6 +14,7 @@ import { RootState } from '../../../types/reduxTypes'
 import { EdgeSceneProps } from '../../../types/routerTypes'
 import { GuiMakeSpendInfo } from '../../../types/types'
 import { getWalletName } from '../../../util/CurrencyWalletHelpers'
+import { getDomainRegInfo } from '../../../util/FioAddressUtils'
 import { SceneWrapper } from '../../common/SceneWrapper'
 import { ButtonsModal } from '../../modals/ButtonsModal'
 import { WalletListModal, WalletListResult } from '../../modals/WalletListModal'
