+++ src/components/scenes/Fio/FioDomainSettingsScene.tsx
@@ -5,12 +5,12 @@ import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
 import { FIO_ADDRESS_DELIMITER } from '../../../constants/WalletAndCurrencyConstants'
 import { formatDate } from '../../../locales/intl'
 import { lstrings } from '../../../locales/strings'
-import { FioActionSubmit } from '../../../modules/FioAddress/components/FioActionSubmit'
-import { getDomainSetVisibilityFee, getRenewalFee, getTransferFee, renewFioDomain, setDomainVisibility } from '../../../modules/FioAddress/util'
 import { connect } from '../../../types/reactRedux'
 import { EdgeSceneProps } from '../../../types/routerTypes'
 import { GuiMakeSpendInfo } from '../../../types/types'
+import { getDomainSetVisibilityFee, getRenewalFee, getTransferFee, renewFioDomain, setDomainVisibility } from '../../../util/FioAddressUtils'
 import { SceneWrapper } from '../../common/SceneWrapper'
+import { FioActionSubmit } from '../../FioAddress/FioActionSubmit'
 import { ButtonsModal } from '../../modals/ButtonsModal'
 import { Airship, showError } from '../../services/AirshipInstance'
 import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
