+++ src/components/scenes/Fio/FioNameConfirmScene.tsx
@@ -4,11 +4,11 @@ import { View } from 'react-native'
 
 import { FIO_ADDRESS_DELIMITER } from '../../../constants/WalletAndCurrencyConstants'
 import { lstrings } from '../../../locales/strings'
-import { FioActionSubmit } from '../../../modules/FioAddress/components/FioActionSubmit'
-import { fioMakeSpend, fioSignAndBroadcast } from '../../../modules/FioAddress/util'
 import { connect } from '../../../types/reactRedux'
 import { EdgeSceneProps } from '../../../types/routerTypes'
+import { fioMakeSpend, fioSignAndBroadcast } from '../../../util/FioAddressUtils'
 import { SceneWrapper } from '../../common/SceneWrapper'
+import { FioActionSubmit } from '../../FioAddress/FioActionSubmit'
 import { ButtonsModal } from '../../modals/ButtonsModal'
 import { Airship, showError } from '../../services/AirshipInstance'
 import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
