+++ src/components/scenes/SendScene.tsx
@@ -20,7 +20,6 @@ import { FioSenderInfo, sendConfirmationUpdateTx, signBroadcastAndSave } from '.
 import { selectWalletToken } from '../../actions/WalletActions'
 import { FIO_STR, getSpecialCurrencyInfo } from '../../constants/WalletAndCurrencyConstants'
 import { lstrings } from '../../locales/strings'
-import { checkRecordSendFee, FIO_NO_BUNDLED_ERR_CODE } from '../../modules/FioAddress/util'
 import { getDisplayDenominationFromState, getExchangeDenominationFromState } from '../../selectors/DenominationSelectors'
 import { config } from '../../theme/appConfig'
 import { connect } from '../../types/reactRedux'
@@ -28,6 +27,7 @@ import { EdgeSceneProps, NavigationBase } from '../../types/routerTypes'
 import { GuiExchangeRates, GuiMakeSpendInfo } from '../../types/types'
 import { getTokenId } from '../../util/CurrencyInfoHelpers'
 import { getWalletName } from '../../util/CurrencyWalletHelpers'
+import { checkRecordSendFee, FIO_NO_BUNDLED_ERR_CODE } from '../../util/FioAddressUtils'
 import { convertTransactionFeeToDisplayFee } from '../../util/utils'
 import { WarningCard } from '../cards/WarningCard'
 import { SceneWrapper } from '../common/SceneWrapper'
