+++ src/components/tiles/AddressTile.tsx
@@ -11,12 +11,12 @@ import { launchPaymentProto } from '../../actions/PaymentProtoActions'
 import { addressWarnings } from '../../actions/ScanActions'
 import { ENS_DOMAINS } from '../../constants/WalletAndCurrencyConstants'
 import { lstrings } from '../../locales/strings'
-import { checkPubAddress } from '../../modules/FioAddress/util'
 import { PaymentProtoError } from '../../types/PaymentProtoError'
 import { connect } from '../../types/reactRedux'
 import { NavigationBase } from '../../types/routerTypes'
 import { GuiMakeSpendInfo } from '../../types/types'
 import { parseDeepLink } from '../../util/DeepLinkParser'
+import { checkPubAddress } from '../../util/FioAddressUtils'
 import { AddressModal } from '../modals/AddressModal'
 import { ScanModal } from '../modals/ScanModal'
 import { Airship, showError } from '../services/AirshipInstance'
