+++ src/util/FioAddressUtils.ts
@@ -3,12 +3,12 @@ import { Disklet } from 'disklet'
 import { EdgeAccount, EdgeCurrencyConfig, EdgeCurrencyWallet, EdgeDenomination, EdgeSpendInfo, EdgeTransaction } from 'edge-core-js'
 import { sprintf } from 'sprintf-js'
 
-import { FIO_STR, getSpecialCurrencyInfo, SPECIAL_CURRENCY_INFO } from '../../constants/WalletAndCurrencyConstants'
-import { lstrings } from '../../locales/strings'
-import { CcWalletMap } from '../../reducers/FioReducer'
-import { BooleanMap, FioAddress, FioConnectionWalletItem, FioDomain, FioObtRecord, StringMap } from '../../types/types'
-import { getWalletName } from '../../util/CurrencyWalletHelpers'
-import { DECIMAL_PRECISION, truncateDecimals } from '../../util/utils'
+import { FIO_STR, getSpecialCurrencyInfo, SPECIAL_CURRENCY_INFO } from '../constants/WalletAndCurrencyConstants'
+import { lstrings } from '../locales/strings'
+import { CcWalletMap } from '../reducers/FioReducer'
+import { BooleanMap, FioAddress, FioConnectionWalletItem, FioDomain, FioObtRecord, StringMap } from '../types/types'
+import { getWalletName } from './CurrencyWalletHelpers'
+import { DECIMAL_PRECISION, truncateDecimals } from './utils'
 
 const CONNECTED_WALLETS = 'ConnectedWallets.json'
 const FIO_ADDRESS_CACHE = 'FioAddressCache.json'
